refactor(core): export `debounce`, `hasFiles` & `objectToFormData`
