+++ packages/core/src/index.ts
@@ -3,6 +3,8 @@ import { Router } from './router'
 export { default as createHeadManager } from './head'
 export { default as shouldIntercept } from './shouldIntercept'
 export { hrefToUrl, mergeDataIntoQueryString, urlWithoutHash } from './url'
+export { hasFiles, objectToFormData } from './form-data'
+export { default as debounce } from './debounce'
 export { resolvePageComponent } from './helper'
 
 export type { Router }
